requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.29
pydantic-settings>=2.2
sqlalchemy[asyncio]>=2.0
asyncpg>=0.29
alembic>=1.13
bcrypt>=4.0
python-jose[cryptography]>=3.3
kubernetes>=29.0
apscheduler>=3.10
httpx>=0.27
pytest>=8.0
pytest-asyncio>=0.23
aiosqlite>=0.20
